# api/openapi.yaml
openapi: 3.0.3
info:
  title: Gold Price Analysis API
  version: 0.1.0
  description: API specification for the gold price analysis project.
servers:
  - url: /api/v1
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
  /dashboard/overview:
    get:
      summary: Get dashboard overview
      responses:
        "200":
          description: OK
  /prices/realtime:
    get:
      summary: Get realtime gold price
      description: Returns the latest normalized gold price in CNY per gram.
      responses:
        "200":
          description: OK
  /prices/history:
    get:
      summary: Get historical gold price candles
      description: Returns candles aggregated from normalized CNY/g ticks.
      parameters:
        - in: query
          name: range
          required: true
          schema:
            type: string
            enum: [1d, 7d, 30d, 90d, 1y]
        - in: query
          name: interval
          required: false
          schema:
            type: string
            enum: [1m, 5m, 15m, 1h, 1d]
      responses:
        "200":
          description: OK
  /prices/stream:
    get:
      summary: Stream realtime gold price
      description: SSE stream that emits `price_tick` and `price_status` events.
      responses:
        "200":
          description: text/event-stream
  /news:
    get:
      summary: Get news list
      description: Returns paginated news articles with category, region and factor filters.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: region
          schema:
            type: string
        - in: query
          name: importance
          schema:
            type: integer
        - in: query
          name: factor_code
          schema:
            type: string
      responses:
        "200":
          description: OK
  /news/{id}:
    get:
      summary: Get news detail
      description: Returns a single news article including full content.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /factors/latest:
    get:
      summary: Get latest factor panel
      description: Returns the latest persisted snapshot for each built-in gold driver factor.
      responses:
        "200":
          description: OK
  /factors/history:
    get:
      summary: Get factor history
      description: Returns historical factor snapshots in ascending time order.
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: range
          required: true
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
      responses:
        "200":
          description: OK
  /factors/definitions:
    get:
      summary: Get factor definitions
      description: Returns the metadata for the 10 built-in gold factors.
      responses:
        "200":
          description: OK
  /reports/latest:
    get:
      summary: Get latest AI report
      description: Returns the latest persisted daily analysis report summary.
      responses:
        "200":
          description: OK
  /reports:
    get:
      summary: Get report list
      description: Returns paginated persisted analysis reports.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
  /reports/{id}:
    get:
      summary: Get report detail
      description: Returns full report content, structured predictions and score detail.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /reports/accuracy/curve:
    get:
      summary: Get report accuracy curve
      description: Returns persisted historical score curve for the selected window.
      parameters:
        - in: query
          name: range
          required: true
          schema:
            type: string
            enum: [30d, 90d, 180d, 1y]
      responses:
        "200":
          description: OK
  /admin/jobs/collect-price:
    post:
      summary: Trigger price collection
      responses:
        "200":
          description: OK
  /admin/jobs/fetch-news:
    post:
      summary: Trigger news collection
      description: Fetches news, applies deduplication and stores normalized articles.
      responses:
        "200":
          description: OK
  /admin/jobs/update-factors:
    post:
      summary: Trigger factor update
      description: Builds the latest factor snapshots from persisted price, news and rule-based signals.
      responses:
        "200":
          description: OK
  /admin/jobs/generate-report:
    post:
      summary: Trigger report generation
      description: Generates a persisted daily report. `report_date` is optional and defaults to today.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                report_date:
                  type: string
                  format: date
      responses:
        "200":
          description: OK
  /admin/jobs/score-report:
    post:
      summary: Trigger report scoring
      description: Scores a persisted daily report. `report_date` is optional and defaults to yesterday.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                report_date:
                  type: string
                  format: date
      responses:
        "200":
          description: OK
  /admin/jobs/definitions:
    get:
      summary: Get job definitions
      description: Returns persisted job schedule definitions, retry policy and latest run status.
      responses:
        "200":
          description: OK
  /admin/jobs/runs:
    get:
      summary: Get job run records
      description: Returns recent manual and scheduled job runs, including retry metadata.
      responses:
        "200":
          description: OK